% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unknown.R
\name{calculateRelate}
\alias{calculateRelate}
\title{this function calculate pairwise estimate between all individuals in the dataset
This function takes in the etaik file, pkla file and indiv file}
\usage{
calculateRelate(hs, etaik, indiv)
}
\arguments{
\item{hs}{.STR dataset}

\item{etaik}{A data frame containing information about subpopulations; output from MULTICLUST or STRUCTURE}

\item{pkla}{A data frame containing information about allele frequency data; output from MULTICLUST or STRUCTURE}
}
\value{
pairwise estimation between each individuals.
}
\description{
this function calculate pairwise estimate between all individuals in the dataset
This function takes in the etaik file, pkla file and indiv file
}
\examples{
calculateRelate(example.str, example.pkla, example.indivq)
}
