% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unknown.R
\name{getLoci}
\alias{getLoci}
\title{Get the number of loci in a data set
This function takes a data set as input and returns the number of loci
in the data set.}
\usage{
getLoci(hs)
}
\arguments{
\item{hs}{A data frame with genetic dat aka the STR file}
}
\value{
The number of loci in the data set.
}
\description{
Get the number of loci in a data set
This function takes a data set as input and returns the number of loci
in the data set.
}
\examples{
data(str)
getLoci(str)
}
