% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InRelate.R
\name{getSubpop}
\alias{getSubpop}
\title{Get number of subpopulations in a dataset
This function takes in a dataset and returns the number of subpopulations}
\usage{
getSubpop(etaik)
}
\arguments{
\item{etaik}{A dataset containing information on subpopulations generated from MULTICLUST or STRUCTURE}
}
\value{
The number of subpopulations in the dataset
}
\description{
Get number of subpopulations in a dataset
This function takes in a dataset and returns the number of subpopulations
}
