% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unknown.R
\name{read_data}
\alias{read_data}
\title{Read in and extract values from various data files
This function reads in three data files and extracts various values such as the number of loci and subpopulations.}
\usage{
read_data(hs, etaik, pkla)
}
\arguments{
\item{hs}{.STR dataset}

\item{etaik}{A data frame containing information about subpopulations; output from MULTICLUST or STRUCTURE}

\item{pkla}{A data frame containing information about allele frequency data; output from MULTICLUST or STRUCTURE}
}
\value{
A list containing data frames for subpopulations, and number of loci etc.
}
\description{
Read in and extract values from various data files
This function reads in three data files and extracts various values such as the number of loci and subpopulations.
}
