% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InRelate.R
\name{readPkla}
\alias{readPkla}
\title{Read in PKLA data
This function takes a data frame containing PKLA data and returns a data frame with only the PKLA data, excluding the first two columns.}
\usage{
readPkla(pkla)
}
\arguments{
\item{pkla}{A data frame containing PKLA data generated from MULTICLUST or STRUCTURE}
}
\value{
A data frame with only the PKLA data.
}
\description{
Read in PKLA data
This function takes a data frame containing PKLA data and returns a data frame with only the PKLA data, excluding the first two columns.
}
