% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InRelate.R
\name{readStr}
\alias{readStr}
\title{Read in STR data
This function reads in the .STR dataset}
\usage{
readStr(hs)
}
\arguments{
\item{hs}{A matrix of haplotype data.}
}
\value{
A matrix of haplotype data.
}
\description{
Read in STR data
This function reads in the .STR dataset
}
