% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InRelate.R
\name{calculateRelate}
\alias{calculateRelate}
\title{This function calculates pairwise estimates of relatedness between all individuals in the dataset.}
\usage{
calculateRelate(hs, etaik, pkla)
}
\arguments{
\item{hs}{A data frame containing STR data.}

\item{etaik}{A data frame containing information about subpopulations; output from MULTICLUST or STRUCTURE.}

\item{pkla}{A data frame containing information about allele frequency data; output from MULTICLUST or STRUCTURE.}
}
\value{
Pairwise estimates of relatedness between each pair of individuals.
}
\description{
This function calculates pairwise estimates of relatedness between all individuals in the dataset.
}
