% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InRelate.R
\name{readIndivq}
\alias{readIndivq}
\title{Read INDIVQ file}
\usage{
readIndivq(etaik)
}
\arguments{
\item{etaik}{An .indivq dataset generated from MULTICLUST or STRUCTURE}
}
\value{
The read in file of the dataset
}
\description{
Read in the dataset from .indivq file
}
