% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InRelate.R
\name{upload_data}
\alias{upload_data}
\title{This function reads in three files, including a STR file, an Indivq file, and a Pkla file, and returns a list containing the data from each file.}
\usage{
upload_data()
}
\value{
A list containing the data from each of the three files.
}
\description{
This function reads in three files, including a STR file, an Indivq file, and a Pkla file, and returns a list containing the data from each file.
}
\seealso{
\code{\link{readStr}}, \code{\link{readIndivq}}, \code{\link{readPkla}}
}
